requirements.txt
discord.py==2.5.2
keyboard==0.13.5
pyautogui==0.9.54
PyDirectInput==1.0.4
pynput==1.8.0
Requests==2.32.3
